Us 12607 : new changes related to timeout pop on un-auth confirmation screen (#427)
* Changed timeout popup for confirmation page

* feat(US-12607 : added redirection for Exit the page link)

* feat()

* feat(US-12607 : included few changes for set page title on static pages)

* feat(US-12607 : merged main to current branch)

* feat(US-12607 : added welsh content and fixed close container issue)

* feat(US-12607 : fixed issues around confirmation screen timeout pop up)

* feat(US-12607 : updated necessary comments in timeout util)

* feat(US-12607 : added claimlist api and isauth properties in stay signed in function)

---------

+++ src/components/AppComponents/TimeoutPopup/timeOutUtils.tsx
@@ -20,7 +20,7 @@ export function clearTimer() {
   clearTimeout(signoutTimeout);
 }
 
-export const initTimeout = (showTimeoutModal, deleteData, isAuthorised) => {
+export const initTimeout = (showTimeoutModal, deleteData, isAuthorised, isConfirmationPage) => {
   // TODO - isAuthorised to be replaced by caseType from pega
   // Fetches timeout length config
   settingTimer();
@@ -32,12 +32,14 @@ export const initTimeout = (showTimeoutModal, deleteData, isAuthorised) => {
     // TODO - unauth and sessiontimeout functionality to be implemented
     showTimeoutModal(true);
     signoutTimeout = setTimeout(() => {
-      if (isAuthorised) {
-        logout();
-      } else {
+      if (!isAuthorised && !isConfirmationPage) {
+        // if the journey is not authorized or from confirmation page , the claim data gets deleted
         deleteData();
         clearTimer();
-        // session ends and deleteData() (pega)
+      } else {
+        // the logout case executes when entire timeout occurs after confirmation page or user clicks
+        // exit survey link in pop after confirmation page
+        logout();
       }
     }, milisecondsTilSignout);
   }, milisecondsTilWarning);
@@ -49,12 +51,13 @@ export function staySignedIn(
   claimsListApi,
   deleteData = null,
   isAuthorised = false,
-  refreshSignin = true
+  refreshSignin = true,
+  isConfirmationPage = false
 ) {
-  if (refreshSignin) {
+  if (refreshSignin && !!claimsListApi) {
     // @ts-ignore
     PCore.getDataPageUtils().getDataAsync(claimsListApi, 'root');
   }
   setShowTimeoutModal(false);
-  initTimeout(setShowTimeoutModal, deleteData, isAuthorised);
+  initTimeout(setShowTimeoutModal, deleteData, isAuthorised, isConfirmationPage);
 }
